+++ cmn-data-svc-server/src/main/java/org/acumos/cds/controller/StepResultController.java
@@ -20,7 +20,7 @@
 package org.acumos.cds.controller;
 
 import java.lang.invoke.MethodHandles;
-import java.util.List;
+import java.util.HashMap;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletResponse;
@@ -43,7 +43,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -92,34 +91,76 @@ public class StepResultController extends AbstractController {
 		return sr;
 	}
 
-	@ApiOperation(value = "Searches for entities with attribute values matching the field name - field value pairs specified as query parameters. " //
-			+ "Defaults to and (conjunction); send junction query parameter '_j=o' for or (disjunction).", //
+	/*
+	 * This method was an early attempt to provide a search feature. Originally
+	 * written with a generic map request parameter to avoid binding field names,
+	 * but that is not supported by Swagger web UI. Now allows use from that web UI
+	 * at the cost of hard-coding many class field names.
+	 */
+	private static final String trackingIdField = "trackingId";
+	private static final String stepCodeField = "stepCode";
+	private static final String solutionIdField = "solutionId";
+	private static final String revisionIdField = "revisionId";
+	private static final String artifactIdField = "artifiactId";
+	private static final String userIdField = "userId";
+	private static final String statusCodeField = "statusCode";
+	private static final String nameField = "name";
+
+	@ApiOperation(value = "Searches for requests with attributes matching the values specified as query parameters. " //
+			+ "Defaults to match all (conjunction); send junction query parameter '_j=o' to match any (disjunction).", //
 			response = MLPPublishRequest.class, responseContainer = "Page")
 	@ApiPageable
 	@ApiResponses({ @ApiResponse(code = 400, message = "Bad request", response = ErrorTransport.class) })
 	@RequestMapping(value = "/" + CCDSConstants.SEARCH_PATH, method = RequestMethod.GET)
 	@ResponseBody
-	public Object searchStepResults(
-			// This actually IS required; set flag to false for swagger UI
-			@ApiParam(value = "Field name - field value pairs as request parameters in the format name=value, minimum 1; repeats allowed. " //
-					+ "Not supported by Swagger web UI.", allowMultiple = true, type = "Array[string]", required = false) //
-			@RequestParam MultiValueMap<String, String> queryParameters, HttpServletResponse response,
-			Pageable pageRequest) {
-		logger.info("searchStepResults {}", queryParameters);
-		cleanPageableParameters(queryParameters);
-		List<String> junction = queryParameters.remove(CCDSConstants.JUNCTION_QUERY_PARAM);
-		boolean isOr = junction != null && junction.size() == 1 && "o".equals(junction.get(0));
+	public Object searchStepResults(@ApiParam(value = "Junction", allowableValues = "a,o") //
+	@RequestParam(name = CCDSConstants.JUNCTION_QUERY_PARAM, required = false) String junction, //
+			@ApiParam(value = "Tracking ID") //
+			@RequestParam(name = trackingIdField, required = false) String trackingId, //
+			@ApiParam(value = "Step code") //
+			@RequestParam(name = stepCodeField, required = false) String stepCode, //
+			@ApiParam(value = "Solution ID") //
+			@RequestParam(name = solutionIdField, required = false) String solutionId, //
+			@ApiParam(value = "Revision ID") //
+			@RequestParam(name = revisionIdField, required = false) String revisionId, //
+			@ApiParam(value = "Artifact ID") //
+			@RequestParam(name = artifactIdField, required = false) String artifactId, //
+			@ApiParam(value = "User ID") //
+			@RequestParam(name = userIdField, required = false) String userId, //
+			@ApiParam(value = "Name") //
+			@RequestParam(name = nameField, required = false) String name, //
+			@ApiParam(value = "Status code") //
+			@RequestParam(name = statusCodeField, required = false) String statusCode, //
+			Pageable pageRequest, HttpServletResponse response) {
+		logger.info("searchStepResults enter");
+		boolean isOr = junction != null && "o".equals(junction);
+		Map<String, Object> queryParameters = new HashMap<>();
+		if (trackingId != null)
+			queryParameters.put(trackingIdField, trackingId);
+		if (stepCode != null)
+			queryParameters.put(stepCodeField, stepCode);
+		if (solutionId != null)
+			queryParameters.put(solutionIdField, solutionId);
+		if (revisionId != null)
+			queryParameters.put(revisionIdField, revisionId);
+		if (artifactId != null)
+			queryParameters.put(artifactIdField, artifactId);
+		if (userId != null)
+			queryParameters.put(userIdField, userId);
+		if (name != null)
+			queryParameters.put(nameField, name);
+		if (statusCode != null)
+			queryParameters.put(statusCodeField, statusCode);
 		if (queryParameters.size() == 0) {
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST, "Missing query", null);
 		}
 		try {
-			Map<String, Object> convertedQryParm = convertQueryParameters(MLPStepResult.class, queryParameters);
-			return stepResultSearchService.findStepResults(convertedQryParm, isOr, pageRequest);
+			return stepResultSearchService.findStepResults(queryParameters, isOr, pageRequest);
 		} catch (Exception ex) {
-			logger.warn("searchStepResults failed: {}", ex.toString());
-			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST,
+			logger.error("searchStepResults failed: {}", ex.toString());
+			response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+			return new ErrorTransport(HttpServletResponse.SC_INTERNAL_SERVER_ERROR,
 					ex.getCause() != null ? ex.getCause().getMessage() : "searchStepResults failed", ex);
 		}
 	}